refactor(messaging)
