+++ tests/test_core_unit.py
@@ -49,54 +49,3 @@ def test_update_drone_position_persists_new_coordinates_and_refreshes_ttl() -> N
         assert redis_client.expirations[drone_key] == 7200
 
     asyncio.run(scenario())
-
-
-def test_resolve_position_request_accepts_transport_metadata_and_returns_position() -> None:
-    async def scenario() -> None:
-        redis_client = FakeRedis()
-        drone_key = state.get_drone_state_key("drone_001")
-        await redis_client.hset(
-            drone_key,
-            mapping=state.serialize_state(
-                state.build_home_state(
-                    {
-                        "drone_id": "drone_001",
-                        "home_lat": 59.9386,
-                        "home_lon": 30.3141,
-                        "home_alt": 120.0,
-                    }
-                )
-            ),
-        )
-
-        response, reason = await core.resolve_position_request(
-            redis_client,
-            {
-                "drone_id": "drone_001",
-                "correlation_id": "req-123",
-                "reply_to": "custom-topic",
-            },
-        )
-
-        assert reason == ""
-        assert response == {
-            "lat": 59.9386,
-            "lon": 30.3141,
-            "alt": 120.0,
-        }
-
-    asyncio.run(scenario())
-
-
-def test_resolve_position_request_returns_error_for_unknown_drone() -> None:
-    async def scenario() -> None:
-        redis_client = FakeRedis()
-        response, reason = await core.resolve_position_request(
-            redis_client,
-            {"drone_id": "drone_404"},
-        )
-
-        assert response is None
-        assert "state not found" in reason
-
-    asyncio.run(scenario())
